create survey is a controller helper
