--- rule/test/.eslintrc.yml
@@ -1,6 +1,0 @@
-rules:
-    max-nested-callbacks: off
-    id-length:
-        - error
-        - exceptions:
-            - t
